+++ src/pages/manageBatch/View_Monthly_Batch.tsx
@@ -1,13 +1,13 @@
 import React, {Component} from 'react';
-import {ONSPanel} from '../components/ONS_DesignSystem/ONSPanel';
-import {ONSButton} from '../components/ONS_DesignSystem/ONSButton';
-import {getMonth, monthNumberToString, qList, toUpperCaseFirstChar} from '../utilities/Common_Functions';
-import {ONSMetadata} from '../components/ONS_DesignSystem/ONSMetadata';
-import {GenericNotFound} from "./GenericNotFound";
+import {ONSPanel} from '../../components/ONS_DesignSystem/ONSPanel';
+import {ONSButton} from '../../components/ONS_DesignSystem/ONSButton';
+import {getMonth, monthNumberToString, qList, toUpperCaseFirstChar} from '../../utilities/Common_Functions';
+import {ONSMetadata} from '../../components/ONS_DesignSystem/ONSMetadata';
+import {GenericNotFound} from "../GenericNotFound";
 import DocumentTitle from "react-document-title";
-import {SurveyAuditModal} from "../components/SurveyAuditModal";
+import {SurveyAuditModal} from "../../components/SurveyAuditModal";
 import {MonthlyBatchUploadTable} from "./MonthlyBatchUploadTable";
-import {getBatchData} from "../utilities/http";
+import {getBatchData} from "../../utilities/http";
 
 interface State {
     UploadsData: Data | null
